Move mime.js to utils, remove libs folder
